Changed mat size constants to size_t

+++ include/mat.h
@@ -12,9 +12,9 @@ namespace linalg
 		vec<Rows, T> cols[Cols];
 
 	public:
-		const static int num_rows = Rows;
-		const static int num_cols = Cols;
-		const static int num_elements = num_rows * num_cols;
+		const static size_t num_rows = Rows;
+		const static size_t num_cols = Cols;
+		const static size_t num_elements = num_rows * num_cols;
 
 		mat(std::initializer_list<std::initializer_list<T>> rows)
 		{
